feat: seperate models into another file
